S922X - enable (optional) Panfrost builds

+++ packages/graphics/mesa/package.mk
@@ -16,11 +16,18 @@ case ${DEVICE} in
 	PKG_URL="${PKG_SITE}.git"
 	PKG_GIT_CLONE_BRANCH="csf"
   ;;
-  RK3*) #Using upstream dev for panfrost
-	PKG_VERSION="db29984c254f60f5daeec0ea4e6048b6ee7902f8"
-	PKG_SITE="https://gitlab.freedesktop.org/mesa/mesa"
-	PKG_URL="${PKG_SITE}.git"
-	PKG_PATCH_DIRS+=" panfrost"
+  RK3*|S922X)
+    if [ "${DEVICE}" = "S922X" -a "${USE_MALI}" != "no" ]; then
+      PKG_VERSION="24.0.4"
+	    PKG_SITE="http://www.mesa3d.org/"
+	    PKG_URL="https://gitlab.freedesktop.org/mesa/mesa/-/archive/mesa-${PKG_VERSION}/mesa-mesa-${PKG_VERSION}.tar.gz"
+    else
+      #Using upstream dev for panfrost
+	    PKG_VERSION="db29984c254f60f5daeec0ea4e6048b6ee7902f8"
+	    PKG_SITE="https://gitlab.freedesktop.org/mesa/mesa"
+	    PKG_URL="${PKG_SITE}.git"
+	    PKG_PATCH_DIRS+=" panfrost"
+    fi
   ;;
   *)
 	PKG_VERSION="24.0.4"
@@ -106,9 +113,7 @@ else
 fi
 
 post_makeinstall_target() {
-  case ${DEVICE} in
-    S922X)
-      rm -f ${INSTALL}/usr/lib/libvulkan_panfrost.so ${INSTALL}/usr/share/vulkan/icd.d/panfrost_icd.aarch64.json
-    ;;
-  esac
+  if [ "${DEVICE}" = "S922X" -a "${USE_MALI}" != "no" ]; then
+    rm -f ${INSTALL}/usr/lib/libvulkan_panfrost.so ${INSTALL}/usr/share/vulkan/icd.d/panfrost_icd.aarch64.json
+  fi
 }
